+++ CHANGELOG.md
@@ -7,7 +7,7 @@ _the graphics engine_
 #### [0.5.0] graphics library (_00 ago 22_)
 
 - **added** - glfw and glad extensions
-- **added** - configuration for enabling the vulkan validation layers
+- **added** - vulkan instance, debug and gpu implementation
 
 ---